Refactor Acceptance Tests This commit: splits acceptancesuice.js into its constituent tests uses bespoke experience key sdkacceptancetests instead of rosetest TEST=auto reran acceptance tests
