+++ package.json
@@ -65,7 +65,7 @@
     "docs": "jsdoc -R README.md -d docs/ -r src/",
     "lint": "semistandard",
     "test": "semistandard && jest",
-    "acceptance": "testcafe safari,chrome tests/acceptance/acceptancesuite.js",
+    "acceptance": "testcafe safari,chrome tests/acceptance/acceptancesuite/*.js",
     "size": "size-limit",
     "fix": "semistandard --fix"
   },
