+++ tests/acceptance/fixtures/html/universal.html
@@ -15,7 +15,7 @@
         <script>
             ANSWERS.init({
                 apiKey: 'df4b24f4075800e5e9705090c54c6c13',
-                experienceKey: 'rosetest',
+                experienceKey: 'sdkacceptancetests',
                 businessId: '2287528',
                 experienceVersion: 'PRODUCTION',
                 templateBundle: TemplateBundle.default,
