+++ tests/acceptance/fixtures/html/vertical.html
@@ -16,7 +16,7 @@
         <script>
             ANSWERS.init({
                 apiKey: 'df4b24f4075800e5e9705090c54c6c13',
-                experienceKey: 'rosetest',
+                experienceKey: 'sdkacceptancetests',
                 businessId: '2287528',
                 templateBundle: TemplateBundle.default,
                 experienceVersion: 'PRODUCTION',
